Create htm.code
comments:Important keys to note:
Following the normal HTML 5 default at the beginning, some key codes are used to modify the page like
The div class: This is an element that groups all element in a page together and the class serves as an element that enables that allows one to write a CSS rule to select that class.
Next is the elements that helps modify the table to the desired styles such as the background color, writing and so other functions. There are also functions that helps customize the height and suitable paddings and and styles.
After which there are elements that helps with the informations given on the platform such as the location , address and navigations . 
With all elements intact the page is previewed.
Note when codes are altered it results to change in the code preview.